Copy function from old
